requirements.txt
tensorflow
numpy
scikit-learn
scipy
google-generativeai
Pillow
python-dotenv
opencv-python
librosa
soundfile
SpeechRecognition
pydub